requirement.txt
langchain==0.0.154
pyPDF2==3.0.1
Python-dotenv==1.0.0
Streamlit==1.18.1
faiss-cpu==1.7.4
streamlit-extras
altair<5
openai==0.27.7
